feat(all): simplified endpoint development
BREAKING CHANGE: Unfortunately, working with version 1 of the Cloudflare
bindings was a pain. As users, you were subject to a weird API
structure, strict—and often incorrect—local validation logic, and having
to use a client that was missing most of the API endpoints. For me
(hi!), it was hard to get validation logic accurate, keeping the models
complete, and adding support for the entire API.

This pre-release introduces a ground-up rewrite of the bindings to
address these issues.

* Endpoints and methods are now generated from a simple specification.
* Removal of those pesky model objects and the inaccurate validation
  logic they contained.
* Methods are now directly on the resource objects.
* Faster (and more direct) unit tests using testdouble.
* You can now provide authentication when calling resource methods,
  which overrides the client's authentication, and "User Service Keys"
  are completely supported.