feat: Workers API
This change adds support for the Cloudflare Workers API endpoints,
allowing users to upload and manage their worker scripts and routes.
Two set of endpoints are provided, one for non-Enterprise zones, and
another for Enterprise zones.

Internally, endpoints can now specify an alternative Content-Type for
the request body and disable automatic JSON parsing of the response.

[terinjokes@gmail.com: dropped non-source changes and added commit msg]
Signed-off-by: Terin Stock <terinjokes@gmail.com>